+++ src/hotload_config.rs
@@ -3,11 +3,11 @@ use serde::Deserialize;
 // 2. Cấu hình đọc từ file (Dùng để can thiệp)
 #[derive(Clone, Copy, Debug, Deserialize)]
 pub struct HotLoadConfig {
-    pub w_empty_override: f32,
-    pub w_disorder_override: f32,
-    pub w_snake_override: f32,
-    pub w_merge_override: f32,
-    pub alpha_override: f32,
+    pub w_empty_override: f64,
+    pub w_disorder_override: f64,
+    pub w_snake_override: f64,
+    pub w_merge_override: f64,
+    pub alpha_override: f64,
 }
 
 impl Default for HotLoadConfig {
